Use 16 bits per pixel in the buffer

+++ include/lcdc.h
@@ -118,7 +118,7 @@ struct lcdc_state_t
 	uint8_t window_y;	/*! Window Y coordinate (0 <= windowy <= 143) */
 	uint8_t window_x;	/*! Window X coordinate (7 <= windowx <= 166) */
 
-	uint8_t out[160][144];	/*! Simulated LCD screen buffer */
+	uint16_t out[160][144];	/*! Simulated LCD screen buffer */
 };
